+++ src/Acl.Fs.Core/Service/Decryption/AesGcm/DecryptorBase.cs
@@ -8,6 +8,7 @@ using Acl.Fs.Core.Abstractions.Service.Shared.KeyDerivation;
 using Acl.Fs.Core.Models;
 using Acl.Fs.Core.Service.Decryption.Shared.Buffer;
 using Acl.Fs.Core.Utility;
+using Acl.Fs.Native.Factory;
 using Microsoft.Extensions.Logging;
 using static Acl.Fs.Constant.Cryptography.CryptoConstants;
 
@@ -87,6 +88,8 @@ internal sealed class DecryptorBase(
                 header,
                 metadataBufferSize,
                 cancellationToken);
+
+            ShellNotifierFactory.NotifyPathUpdated(instruction.DestinationPath);
 
             await _auditService.AuditDecryptionCompleted(cancellationToken);
         }