+++ src/Acl.Fs.Core/Service/Encryption/XChaCha20Poly1305/EncryptorBase.cs
@@ -10,6 +10,7 @@ using Acl.Fs.Core.Models;
 using Acl.Fs.Core.Policy;
 using Acl.Fs.Core.Service.Encryption.Shared.Buffer;
 using Acl.Fs.Core.Utility;
+using Acl.Fs.Native.Factory;
 using Microsoft.Extensions.Logging;
 using NSec.Cryptography;
 using static Acl.Fs.Constant.Cryptography.CryptoConstants;
@@ -93,6 +94,8 @@ internal sealed class EncryptorBase(
                 bufferManager,
                 metadataBufferSize,
                 cancellationToken);
+
+            ShellNotifierFactory.NotifyPathUpdated(instruction.DestinationPath);
 
             await _auditService.AuditEncryptionCompleted(cancellationToken);
         }