+++ src/Acl.Fs.Native/Platform/Windows/WindowsConstants.cs
@@ -5,5 +5,6 @@ internal static class WindowsConstants
     internal static class Libraries
     {
         internal const string Kernel32LibraryName = "kernel32.dll";
+        internal const string Shell32LibraryName = "shell32.dll";
     }
 }